+++ src/biped_motion_planner/biped_motion_planner/ss_to_ds_manager.py
@@ -21,8 +21,8 @@ class SSToDSManager:
         self._support_side: SupportSide = "undefined"
         self._stance_func = None
         self._swing_func = None
-        self._xLFOOT_W_norm: Optional[NDArray[np.float64]] = None
-        self._xRFOOT_W_norm: Optional[NDArray[np.float64]] = None
+        self._xLFOOT_W_norm: Optional[NDArray[np.float32]] = None
+        self._xRFOOT_W_norm: Optional[NDArray[np.float32]] = None
         self._validate_parameters()
 
     def set_stance_side(self, side: LegSide) -> None:
@@ -53,12 +53,12 @@ class SSToDSManager:
         if not self._if_side_defined():
             raise ValueError("Stance or swing or support side is undefined.")
         baselink_move_dist_xB: float = self._calc_baselink_move_dist_xB()
-        p_S_baselink_target: NDArray[np.float64] = self._calc_p_S_baselink_target(
+        p_S_baselink_target: NDArray[np.float32] = self._calc_p_S_baselink_target(
             baselink_move_dist_xB, p_W, q_W)
-        p_S_stance: NDArray[np.float64] = self._calc_p_S_stance(p_W, q_W)
-        p_S_swing_end: NDArray[np.float64] = self._calc_p_S_swing_end(
+        p_S_stance: NDArray[np.float32] = self._calc_p_S_stance(p_W, q_W)
+        p_S_swing_end: NDArray[np.float32] = self._calc_p_S_swing_end(
             p_S_baselink_target, p_S_stance)
-        p_W_swing_start: NDArray[np.float64] = self._calc_p_W_swing_start(
+        p_W_swing_start: NDArray[np.float32] = self._calc_p_W_swing_start(
             p_W, q_W)
         # the origin of swing_of_s is at the foot mid, need to shift to foot front(the target point)
         swing_of_s = self._build_swing_of_s(p_W_swing_start, p_S_swing_end)
@@ -73,12 +73,12 @@ class SSToDSManager:
         stance_alpha = float(self._stance_func(s_value))
         return self._compose_stance_joint_pose(stance_alpha)
 
-    def calc_swing_position(self, s_value: float) -> NDArray[np.float64]:
+    def calc_swing_position(self, s_value: float) -> NDArray[np.float32]:
         if self._swing_func is None:
             raise ValueError("Swing trajectory is not yet built.")
         if not (0.0 <= s_value <= 1.0):
             raise ValueError("s_value must be between 0 and 1.")
-        return np.array([float(f(s_value)) for f in self._swing_func], dtype=np.float64)
+        return np.array([float(f(s_value)) for f in self._swing_func], dtype=np.float32)
 
     def clear_phase_state(self) -> None:
         self._stance_side = "undefined"
@@ -100,13 +100,13 @@ class SSToDSManager:
     def _if_side_defined(self) -> bool:
         return self._stance_side in VALID_LEG_SIDES and self._swing_side in VALID_LEG_SIDES and self._support_side in VALID_SUPPORT_SIDES
 
-    def _build_swing_of_s(self, p_W_swing_start: NDArray[np.float64], p_S_swing_end: NDArray[np.float64]) -> NDArray[object]:
+    def _build_swing_of_s(self, p_W_swing_start: NDArray[np.float32], p_S_swing_end: NDArray[np.float32]) -> NDArray[object]:
         # p_S_swing_start and p_S_swing_end are at foot mid, need to shift to foot front(the target point)
         p_S_swing_start = np.array(
-            [p_W_swing_start[0], p_W_swing_start[1], 0.0], dtype=np.float64)
+            [p_W_swing_start[0], p_W_swing_start[1], 0.0], dtype=np.float32)
         p_S_swing_mid = (p_S_swing_start + p_S_swing_end) / 2.0
         p_W_swing_mid = np.array(
-            [p_S_swing_mid[0], p_S_swing_mid[1], Config.SWING_TRAJECTORY_MID_HEIGHT], dtype=np.float64)
+            [p_S_swing_mid[0], p_S_swing_mid[1], Config.SWING_TRAJECTORY_MID_HEIGHT], dtype=np.float32)
 
         raw_swing_of_s = ((1-self._s)**2)*p_W_swing_start + 2*(1-self._s) * \
             self._s*p_W_swing_mid + (self._s**2)*p_S_swing_end  # foot mid
@@ -127,7 +127,7 @@ class SSToDSManager:
     def _calc_baselink_move_dist_xB(self) -> float:
         return (Config.THIGH_LEN + Config.CALF_LEN)*np.sin(np.deg2rad(STANCE_LEG_JOINT_ALPHA))
 
-    def _calc_p_S_baselink_target(self, baselink_move: float, p_W: Mapping[str, Vector3], q_W: Mapping[str, Quaternion]) -> NDArray[np.float64]:
+    def _calc_p_S_baselink_target(self, baselink_move: float, p_W: Mapping[str, Vector3], q_W: Mapping[str, Quaternion]) -> NDArray[np.float32]:
         R_WB = LinearAlgebraUtils.quaternion_to_rotation_matrix(
             q_W["baselink"])
         xB_W = R_WB[:, 0]
@@ -140,52 +140,52 @@ class SSToDSManager:
         ])
         return p_S_baselink + baselink_move * xB_S_norm
 
-    def _calc_p_S_stance(self, p_W: Mapping[str, Vector3], q_W: Mapping[str, Quaternion]) -> NDArray[np.float64]:
+    def _calc_p_S_stance(self, p_W: Mapping[str, Vector3], q_W: Mapping[str, Quaternion]) -> NDArray[np.float32]:
         if self._stance_side == "left":
             self._xLFOOT_W_norm = self._calc_xFOOT_W_norm(q_W["l_foot"])
             p_W_l_foot = np.array(
-                [p_W["l_foot"].x, p_W["l_foot"].y, p_W["l_foot"].z], dtype=np.float64)
+                [p_W["l_foot"].x, p_W["l_foot"].y, p_W["l_foot"].z], dtype=np.float32)
             p_W_l_foot_mid = p_W_l_foot - Config.FOOT_LINK_X_SEMI_LENGTH * self._xLFOOT_W_norm
             p_S_stance = np.array(
-                [p_W_l_foot_mid[0], p_W_l_foot_mid[1], 0.0], dtype=np.float64)
+                [p_W_l_foot_mid[0], p_W_l_foot_mid[1], 0.0], dtype=np.float32)
         elif self._stance_side == "right":
             self._xRFOOT_W_norm = self._calc_xFOOT_W_norm(q_W["r_foot"])
             p_W_r_foot = np.array(
-                [p_W["r_foot"].x, p_W["r_foot"].y, p_W["r_foot"].z], dtype=np.float64)
+                [p_W["r_foot"].x, p_W["r_foot"].y, p_W["r_foot"].z], dtype=np.float32)
             p_W_r_foot_mid = p_W_r_foot - Config.FOOT_LINK_X_SEMI_LENGTH * self._xRFOOT_W_norm
             p_S_stance = np.array(
-                [p_W_r_foot_mid[0], p_W_r_foot_mid[1], 0.0], dtype=np.float64)
+                [p_W_r_foot_mid[0], p_W_r_foot_mid[1], 0.0], dtype=np.float32)
         else:
             raise ValueError("Stance side is undefined.")
         return p_S_stance
 
-    def _calc_xFOOT_W_norm(self, q_W_foot: Quaternion) -> NDArray[np.float64]:
-        R_W_FOOT: NDArray[np.float64] = LinearAlgebraUtils.quaternion_to_rotation_matrix(
+    def _calc_xFOOT_W_norm(self, q_W_foot: Quaternion) -> NDArray[np.float32]:
+        R_W_FOOT: NDArray[np.float32] = LinearAlgebraUtils.quaternion_to_rotation_matrix(
             q_W_foot)
         xFOOT_W = R_W_FOOT[:, 0]
         return LinearAlgebraUtils.normalize_vec(xFOOT_W)
 
-    def _calc_zFOOT_W_norm(self, q_W_foot: Quaternion) -> NDArray[np.float64]:
-        R_W_FOOT: NDArray[np.float64] = LinearAlgebraUtils.quaternion_to_rotation_matrix(
+    def _calc_zFOOT_W_norm(self, q_W_foot: Quaternion) -> NDArray[np.float32]:
+        R_W_FOOT: NDArray[np.float32] = LinearAlgebraUtils.quaternion_to_rotation_matrix(
             q_W_foot)
         zFOOT_W = R_W_FOOT[:, 2]
         return LinearAlgebraUtils.normalize_vec(zFOOT_W)
 
-    def _calc_p_S_swing_end(self, p_S_baselink_target: NDArray[np.float64], p_S_stance: NDArray[np.float64]) -> NDArray[np.float64]:
+    def _calc_p_S_swing_end(self, p_S_baselink_target: NDArray[np.float32], p_S_stance: NDArray[np.float32]) -> NDArray[np.float32]:
         return (p_S_baselink_target - (1-LAMBDA_FOR_SWING_END) * p_S_stance) / LAMBDA_FOR_SWING_END
 
-    def _calc_p_W_swing_start(self, p_W: Mapping[str, Vector3], q_W: Mapping[str, Quaternion]) -> NDArray[np.float64]:
+    def _calc_p_W_swing_start(self, p_W: Mapping[str, Vector3], q_W: Mapping[str, Quaternion]) -> NDArray[np.float32]:
         if self._swing_side == "left":
             self._xLFOOT_W_norm = self._calc_xFOOT_W_norm(q_W["l_foot"])
             p_W_l_foot = np.array(
-                [p_W["l_foot"].x, p_W["l_foot"].y, p_W["l_foot"].z], dtype=np.float64)
+                [p_W["l_foot"].x, p_W["l_foot"].y, p_W["l_foot"].z], dtype=np.float32)
             p_W_l_foot_mid = p_W_l_foot - Config.FOOT_LINK_X_SEMI_LENGTH * self._xLFOOT_W_norm
             zLFOOT_W_norm = self._calc_zFOOT_W_norm(q_W["l_foot"])
             p_W_swing_start = p_W_l_foot_mid - Config.FOOT_LEN*zLFOOT_W_norm
         elif self._swing_side == "right":
             self._xRFOOT_W_norm = self._calc_xFOOT_W_norm(q_W["r_foot"])
             p_W_r_foot = np.array(
-                [p_W["r_foot"].x, p_W["r_foot"].y, p_W["r_foot"].z], dtype=np.float64)
+                [p_W["r_foot"].x, p_W["r_foot"].y, p_W["r_foot"].z], dtype=np.float32)
             p_W_r_foot_mid = p_W_r_foot - Config.FOOT_LINK_X_SEMI_LENGTH * self._xRFOOT_W_norm
             zRFOOT_W_norm = self._calc_zFOOT_W_norm(q_W["r_foot"])
             p_W_swing_start = p_W_r_foot_mid - Config.FOOT_LEN*zRFOOT_W_norm
